i18n-ize moment strings and move poll to fun directory
